moved game board logic to its own class
